+++ automation/src/epilot/models/components/createdocumentactionconfig.py
@@ -33,6 +33,8 @@ class CreateDocumentActionConfig:
     r"""Flag indicating whether the action was created automatically or manually"""
     flow_action_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('flow_action_id'), 'exclude': lambda f: f is None }})
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})
+    is_bulk_action: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('is_bulk_action'), 'exclude': lambda f: f is None }})
+    r"""Flag indicating whether the same action can be in bulk in a single execution. e.g; send-email / map-entity"""
     name: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name'), 'exclude': lambda f: f is None }})
     reason: Optional[CreateDocumentActionConfigReason] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('reason'), 'exclude': lambda f: f is None }})
     type: Optional[CreateDocumentActionConfigType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})